Revert use of build kit
In #11, I tried to find a configuration that works both in Docker 19
and 20. However, this breaks the build.

This reverts the change as there are bugs popping up related on Docker
20 anyway, such as https://github.com/docker/buildx/pull/475

We use Docker from the OS distribution, so unlikely to accidentally
walk into a Docker 20. In other words, when it becomes stable, we can
try again.